-> 6N24
